data_feed_collect/collector.py
"""Main collector module for data feed collection."""

import logging
import os
from typing import Any, Dict, Optional

import yaml


class DataCollector:
    """Main class for collecting data from various sources.

    This class handles data collection from configured sources,
    processing the collected data, and storing it in the specified format.
    """

    def __init__(self, config_path: Optional[str] = None):
        """Initialize the DataCollector.

        Args:
            config_path: Path to the configuration file (YAML format).
        """
        self.logger = logging.getLogger(__name__)
        self.config = {}
        if config_path and os.path.exists(config_path):
            try:
                with open(config_path, "r") as f:
                    self.config = yaml.safe_load(f)
            except Exception as e:
                self.logger.error(f"Failed to load configuration: {e}")

    def collect(self, source_name: str) -> Dict[str, Any]:
        """Collect data from the specified source.

        Args:
            source_name: Name of the source to collect from (defined in config).

        Returns:
            Dict containing the collected data.

        Raises:
            ValueError: If the source is not configured.
        """
        if source_name not in self.config.get("sources", {}):
            raise ValueError(f"Source '{source_name}' not configured")

        # Implementation placeholder - would be expanded in actual use
        self.logger.info(f"Collecting data from {source_name}")
        return {"source": source_name, "data": []}

    def process(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """Process the collected data.

        Args:
            data: Data to process.

        Returns:
            Processed data.
        """
        # Implementation placeholder - would be expanded in actual use
        self.logger.info("Processing data")
        return {"processed": True, **data}

    def save(self, data: Dict[str, Any], output_path: str) -> None:
        """Save the processed data to the specified path.

        Args:
            data: Data to save.
            output_path: Path where to save the data.
        """
        # Implementation placeholder - would be expanded in actual use
        self.logger.info(f"Saving data to {output_path}")
        # In a real implementation, this would save to file, database, etc.