pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data-feed-collect"
version = "0.1.0"
description = "A Python project for collecting and processing data feeds"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "karimQuant", email = "karim.taaam@gmail.com"}
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "click>=8.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0"
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88